refactor: move all marketing CTA components together
