+++ packages/shared/src/components/FeedItemComponent.tsx
@@ -21,8 +21,8 @@ import { useFeedLayout, UseVotePost } from '../hooks';
 import { CollectionCard } from './cards/CollectionCard';
 import { CollectionList } from './cards/list/CollectionList';
 import { AcquisitionFormCard } from './cards/AcquisitionFormCard';
-import { MarketingCtaCard } from './cards';
-import { MarketingCtaList } from './cards/list/MarketingCtaList';
+import { MarketingCtaCard } from './marketingCta';
+import { MarketingCtaList } from './marketingCta/MarketingCtaList';
 import { FeedItemType } from './cards/common';
 import { PublicSquadEligibilityCard } from './squads/PublicSquadEligibilityCard';
 import { FeedSurveyCard } from './cards/survey';
