+++ packages/shared/src/components/cards/index.ts
@@ -1,2 +1,1 @@
 export * from './FeedbackCard';
-export * from './MarketingCta';
